Update the db conn string
We want to be able to dynamically change the DNS records without the
need to create a PR in the DNS repository and wait for review from
another team. This will enable us to act fast in case of emergency.

In this change we update the DB connection string to look up the
`upp.ft.com` DNS records - staging-rds-pac.upp.ft.com and
prod-rds-pac.upp.ft.com